328 making policies

+++ resources/views/pages/animals/⚡show/show.php
@@ -11,12 +11,14 @@ new #[Layout('layouts.app')] class extends Component
 
     public function mount(Animal $animal): void
     {
+        $this->authorize('view', $animal);
         $this->animal = $animal;
     }
 
     public function destroy()
     {
         $animal = Animal::findOrFail($this->animal->id);
+        $this->authorize('delete', $animal);
         $this->animal->delete();
         return redirect(route('pages::animals.index', ['locale' => app()->getLocale()]));
     }
@@ -29,6 +31,8 @@ new #[Layout('layouts.app')] class extends Component
         $age = $today->diff($birthday);
         $age = $age->y;
 
+
         return view('pages.animals.⚡show.show', ['animal'=> $this->animal, 'age'=>$age])->title(__('general.animals_show'));
+
     }
 };
